# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trak fitness coaching platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients/{clientId}/workout-logs:
    get:
      operationId: listWorkoutLogs
      tags: [workouts]
      summary: List workout logs for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workout logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkoutLog"
    post:
      operationId: createWorkoutLog
      tags: [workouts]
      summary: Log a completed workout
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutLogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutLog"
  /clients/{clientId}/workout-logs/{logId}:
    get:
      operationId: getWorkoutLog
      tags: [workouts]
      summary: Get a workout log with set details
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: logId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workout log detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutLogDetail"
    patch:
      operationId: updateWorkoutLog
      tags: [workouts]
      summary: Update a workout log (early exit reason, status, etc.)
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: logId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutLogUpdate"
      responses:
        "200":
          description: Updated workout log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutLog"
  /clients/{clientId}/workout-logs/{logId}/sets:
    post:
      operationId: logSet
      tags: [workouts]
      summary: Log a set within a workout
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: logId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetLogInput"
      responses:
        "201":
          description: Set logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetLog"

  # ── Measurements ──────────────────────────────────────────────────────────
...
  /clients/{clientId}/sleep:
    get:
      operationId: listSleepLogs
      tags: [sleep]
      summary: List sleep logs for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sleep logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SleepLog"
    post:
      operationId: logSleep
      tags: [sleep]
      summary: Log a sleep entry
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SleepLogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepLog"
  /clients/{clientId}/sleep/{sleepId}:
    delete:
      operationId: deleteSleepLog
      tags: [sleep]
      summary: Delete a sleep log
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: sleepId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Nutrition Screenshots ─────────────────────────────────────────────────
...
  /clients/{clientId}/photos:
    get:
      operationId: listProgressPhotos
      tags: [photos]
      summary: List progress photos for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgressPhoto"
    post:
      operationId: createProgressPhoto
      tags: [photos]
      summary: Upload a progress photo entry
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressPhotoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressPhoto"
  /clients/{clientId}/photos/{photoId}:
    delete:
      operationId: deleteProgressPhoto
      tags: [photos]
      summary: Delete a progress photo
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: photoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Assignments ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    WorkoutLog:
      type: object
      required: [id, clientId, date, status, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        programDayId:
          type: ["integer", "null"]
        programDayName:
          type: ["string", "null"]
        date:
          type: string
          format: date
        durationMinutes:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [in_progress, completed, early_exit, cancelled]
        createdAt:
          type: string
          format: date-time
    WorkoutLogInput:
      type: object
      required: [date]
      properties:
        programDayId:
          type: integer
        date:
          type: string
          format: date
        durationMinutes:
          type: integer
        notes:
          type: string
    WorkoutLogUpdate:
      type: object
      properties:
        notes:
          type: string
        status:
          type: string
          enum: [in_progress, completed, early_exit, cancelled]
    WorkoutLogDetail:
      type: object
      required: [id, clientId, date, status, sets]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        programDayId:
          type: ["integer", "null"]
        programDayName:
          type: ["string", "null"]
        date:
          type: string
          format: date
        durationMinutes:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [in_progress, completed, early_exit, cancelled]
        sets:
          type: array
          items:
            $ref: "#/components/schemas/SetLog"
        createdAt:
          type: string
          format: date-time
    SetLog:
      type: object
      required: [id, workoutLogId, exerciseId, exerciseName, setNumber, reps]
      properties:
        id:
          type: integer
        workoutLogId:
          type: integer
        exerciseId:
          type: integer
        exerciseName:
          type: string
        setNumber:
          type: integer
        reps:
          type: integer
        weight:
          type: ["number", "null"]
        weightUnit:
          type: ["string", "null"]
          enum: [lbs, kg, null]
        rpe:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    SetLogInput:
      type: object
      required: [exerciseId, setNumber, reps]
      properties:
        exerciseId:
          type: integer
        setNumber:
          type: integer
        reps:
          type: integer
        weight:
          type: number
        weightUnit:
          type: string
          enum: [lbs, kg]
        rpe:
          type: integer
        notes:
          type: string
...
    SleepLog:
      type: object
      required: [id, clientId, date, hoursSlept, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        date:
          type: string
          format: date
        hoursSlept:
          type: number
        quality:
          type: ["string", "null"]
          enum: [poor, fair, good, great, null]
        energyRating:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    SleepLogInput:
      type: object
      required: [date, hoursSlept]
      properties:
        date:
          type: string
          format: date
        hoursSlept:
          type: number
        quality:
          type: string
          enum: [poor, fair, good, great]
        energyRating:
          type: integer
        notes:
          type: string

    # ── Nutrition ───────────────────────────────────────────────────────────
...
    ProgressPhoto:
      type: object
      required: [id, clientId, date, imageUrl, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        date:
          type: string
          format: date
        imageUrl:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ProgressPhotoInput:
      type: object
      required: [date, imageUrl]
      properties:
        date:
          type: string
          format: date
        imageUrl:
          type: string
        notes:
          type: string

    # ── Assignments ─────────────────────────────────────────────────────────
...